TASK-427-A step-3: ROOT-CAUSE FIX ColpushOps.selfTest javac-constant-fold poison — 'IDS_CAP*ROW_D < (1<<31)' folds to false at COMPILE TIME (1<<31 == Integer.MIN_VALUE in int arithmetic; 6291456 < -2147483648 = false; javac folds whole && chain to iconst_0 stub) => selfTest ALWAYS returned false => colpush hook dormant fail-closed on every REBUILT blob (mobfeed 2ef6765, chunksend 6f92ea7) => pushEntities redirect never active => SoA feed dead (mobSlots=0 x895, DATA-PLAN FAIL, windowLen=0 whole run a426l2). Certified mega-420 legs predate the rust selfTest-gate (no Java selfTest call in register flow) — that's why cmp420_colpush armed there. FIX: long arithmetic (long)IDS_CAP*(long)ROW_D < (1L<<31) = 6291456 < 2147483648 = constant-true, runtime MobPushOps checks stay live in bytecode (javap verified: invokestatic idCount/idCapacity/byIdArr present, no iconst_0). blobs rebuilt (build_419a_blobs_all.sh), check_blobs_sync ALL IN SYNC, cargo check --lib ok

+++ colpush/net/minecraft/world/entity/ColpushOps.java
@@ -406,9 +406,20 @@ public final class ColpushOps {
      * MobPushOps. Вызывается один раз из bulkTick (маркер в stdout).
      */
     public static boolean selfTest() {
+        // TASK-427 ROOT-CAUSE FIX: прежняя версия содержала
+        // `IDS_CAP * ROW_D < (1 << 31)` — в int-арифметике 1<<31 =
+        // Integer.MIN_VALUE, поэтому сравнение 6291456 < -2147483648 ЛОЖНО
+        // ещё на этапе КОМПИЛЯЦИИ; javac сворачивал всю &&-цепочку в
+        // константу false (bytecode: iconst_0) => selfTest ВСЕГДА возвращал
+        // false => colpush hook dormant fail-closed на каждом пересобранном
+        // блобе (mobfeed/chunksend) => SoA-кормление не просыпалось
+        // (mobSlots=0 x895, DATA-PLAN FAIL). Лечение: long-арифметика
+        // (1L<<31 = 2147483648) — конъюнкты становятся константно-истинными,
+        // рантайм-проверки MobPushOps остаются живыми в байткоде.
         boolean ok = ROW_D == 6 && ROW_I == 3 && FLAG_INCLUDE == 1
                 && IDS_CAP == (1 << 20)
-                && IDS_CAP * ROW_D < (1 << 31) && IDS_CAP * ROW_I < (1 << 31)
+                && (long) IDS_CAP * (long) ROW_D < (1L << 31)
+                && (long) IDS_CAP * (long) ROW_I < (1L << 31)
                 && MobPushOps.idCount() >= 0 && MobPushOps.idCapacity() > 0
                 && MobPushOps.byIdArr() != null;
         if (!ok) {
